multishots

+++ Progra3/src/game/logic/ataque/Teletransportacion.java
@@ -1,6 +1,6 @@
 package game.logic.ataque;
 
-import game.logic.Nodo;
+import game.logic.Estructuras.Nodo;
 
 import java.util.List;
 
@@ -13,7 +13,7 @@ public class Teletransportacion extends Ataque {
     public void atacar(int promedioGrafo, List<Nodo> recorrido){
         int randint = random.nextInt(2)+3;
         promedio = promedioGrafo;
-        int herida = (int) (randint * (10* promedioGrafo));
+        int herida = (randint * (10* promedioGrafo));
         atacar(recorrido,herida,FASTWAY);
     }
 }
